extract classes into separate files, add as_list method
